feat: add intervals for recurrence dates
