+++ src/server/api/routers/google.ts
@@ -20,11 +20,9 @@ import {
 	updateAvailabilityEvent,
 	updatePunchData,
 } from "~/lib/google";
-import { logger } from "~/lib/logger";
 import type { Client } from "~/lib/models";
 import type { DuplicateGroup } from "~/lib/types";
 import { getDistanceSQL, getInsuranceShortName } from "~/lib/utils";
-
 import {
 	assertPermission,
 	type Context,
@@ -45,8 +43,6 @@ const availabilitySchema = z.object({
 });
 const CACHE_KEY_PUNCHLIST = "google:sheets:punchlist";
 const CACHE_KEY_MISSING_PUNCHLIST = "google:sheets:missing-punchlist";
-
-const log = logger.child({ module: "GoogleRouter" });
 
 const getPreviewData = async (ctx: Context, clientId: number) => {
 	const client = await ctx.db.query.clients.findFirst({
@@ -454,10 +450,8 @@ export const googleRouter = createTRPCRouter({
 	createAvailability: protectedProcedure
 		.input(availabilitySchema)
 		.mutation(async ({ ctx, input }) => {
-			if (!ctx.session) {
-				throw new TRPCError({
-					code: "UNAUTHORIZED",
-				});
+			if (!ctx.session.user.accessToken || !ctx.session.user.refreshToken) {
+				throw new Error("No access token or refresh token");
 			}
 
 			let summary: string;
@@ -553,7 +547,7 @@ export const googleRouter = createTRPCRouter({
 							masterRecurrenceMap.set(masterId, masterEvent.data.recurrence);
 						}
 					} catch (error) {
-						log.error(
+						ctx.logger.error(
 							{ error, masterId },
 							"Error fetching master event recurrence",
 						);
@@ -661,7 +655,7 @@ export const googleRouter = createTRPCRouter({
 					eventId: event.id,
 				};
 			} catch (error) {
-				log.error({ error, input }, "Error in updateAvailability");
+				ctx.logger.error({ error, input }, "Error in updateAvailability");
 				if (error instanceof TRPCError) throw error;
 				throw new TRPCError({
 					code: "INTERNAL_SERVER_ERROR",
@@ -688,7 +682,7 @@ export const googleRouter = createTRPCRouter({
 					message: "Availability event deleted successfully.",
 				};
 			} catch (error) {
-				log.error({ error, input }, "Error in deleteAvailability");
+				ctx.logger.error({ error, input }, "Error in deleteAvailability");
 				if (error instanceof TRPCError) throw error;
 				throw new TRPCError({
 					code: "INTERNAL_SERVER_ERROR",
